LAB3: lab3 is working now

+++ include/field/conditional_field.h
@@ -11,7 +11,7 @@ template <typename LetterType, int N>
 class ConditionalField : Field<LetterType, N> {
     std::function<bool(const Vector&)> filter;
 
-	std::shared_ptr<Vector> _try_to_generate_element(int i) const {
+	std::shared_ptr<Vector> _try_to_generate_element(int64_t i) const {
         int alphabet_power = this->alphabet.size();
         std::vector<int> translated = this->translate_int(i, alphabet_power);
 
@@ -26,9 +26,9 @@ public:
         : Field<Vector, N>(alphabet, enable_caching), filter(filter)
          {}
 
-	int size() = delete;
+	int64_t size() = delete;
 
-	Vector& generate_element(int i) const {
+	Vector& generate_element(int64_t i) const {
 		std::shared_ptr<Vector> result;
 		do {
 			result = Field<T, N>::generate_element(i++);
